sch_dualpi2: Run prob update timer in softirq to avoid deadlock
Dualpi2 qdisc automatically sets a timer, and it will run constantly
to update the probability. This timer acquires the qdisc root lock
and runs in hardirq, while the qdisc root lock is also acquired
whenever a packet arrives for this qdisc.

Since the dualpi2 timer callback runs in hardirq, it may interrupt
the packet processing running in softirq. And if they run on the same
CPU, it will acquire the same lock and cause a deadlock.

Signed-off-by: Chia-Yu Chang <chia-yu.chang@nokia-bell-labs.com>